+++ Swamp/Messages/RPC/Callee/RegisterSwampMessage.swift
@@ -11,11 +11,11 @@ import Foundation
 /// [Register, requestId|number, options|dict, proc|string]
 class RegisterSwampMessage: SwampMessage {
     
-    let requestId: Int
+    let requestId: Int64
     let options: [String: AnyObject]
     let proc: String
     
-    init(requestId: Int, options: [String: AnyObject], proc: String) {
+    init(requestId: Int64, options: [String: AnyObject], proc: String) {
         self.requestId = requestId
         self.options = options
         self.proc = proc
@@ -23,7 +23,7 @@ class RegisterSwampMessage: SwampMessage {
     
     // MARK: SwampMessage protocol
     required init(payload: [Any]) {
-        self.requestId = payload[0] as! Int
+        self.requestId = payload[0] as! Int64
         self.options = payload[1] as! [String: AnyObject]
         self.proc = payload[2] as! String
     }
