+++ Swamp/SwampSession.swift
@@ -27,11 +27,11 @@ public typealias ErrorPublishCallback = (_ details: [String: Any], _ error: Stri
 // TODO: Expose only an interface (like Cancellable) to user
 open class Subscription {
     fileprivate let session: SwampSession
-    internal let subscription: Int
+    internal let subscription: Int64
     internal let eventCallback: EventCallback
     fileprivate var isActive: Bool = true
 
-    internal init(session: SwampSession, subscription: Int, onEvent: @escaping EventCallback) {
+    internal init(session: SwampSession, subscription: Int64, onEvent: @escaping EventCallback) {
         self.session = session
         self.subscription = subscription
         self.eventCallback = onEvent
@@ -56,7 +56,7 @@ open class Subscription {
 
 public protocol SwampSessionDelegate {
     func swampSessionHandleChallenge(_ authMethod: String, extra: [String: Any]) -> String
-    func swampSessionConnected(_ session: SwampSession, sessionId: Int)
+    func swampSessionConnected(_ session: SwampSession, sessionId: Int64)
     func swampSessionEnded(_ reason: String)
 }
 
@@ -80,28 +80,28 @@ open class SwampSession: SwampTransportDelegate {
     fileprivate let authextra: [String: Any]?
 
     // MARK: State members
-    fileprivate var currRequestId: Int = 1
+    fileprivate var currRequestId: Int64 = 1
 
     // MARK: Session state
     fileprivate var serializer: SwampSerializer?
-    fileprivate var sessionId: Int?
+    fileprivate var sessionId: Int64?
     fileprivate var routerSupportedRoles: [SwampRole]?
 
     // MARK: Call role
     //                         requestId
-    fileprivate var callRequests: [Int: (callback: CallCallback, errorCallback: ErrorCallCallback)] = [:]
+    fileprivate var callRequests: [Int64: (callback: CallCallback, errorCallback: ErrorCallCallback)] = [:]
 
     // MARK: Subscriber role
     //                              requestId
-    fileprivate var subscribeRequests: [Int: (callback: SubscribeCallback, errorCallback: ErrorSubscribeCallback, eventCallback: EventCallback)] = [:]
+    fileprivate var subscribeRequests: [Int64: (callback: SubscribeCallback, errorCallback: ErrorSubscribeCallback, eventCallback: EventCallback)] = [:]
     //                          subscription
-    fileprivate var subscriptions: [Int: Subscription] = [:]
+    fileprivate var subscriptions: [Int64: Subscription] = [:]
     //                                requestId
-    fileprivate var unsubscribeRequests: [Int: (subscription: Int, callback: UnsubscribeCallback, errorCallback: ErrorUnsubscribeCallback)] = [:]
+    fileprivate var unsubscribeRequests: [Int64: (subscription: Int64, callback: UnsubscribeCallback, errorCallback: ErrorUnsubscribeCallback)] = [:]
 
     // MARK: Publisher role
     //                            requestId
-    fileprivate var publishRequests: [Int: (callback: PublishCallback, errorCallback: ErrorPublishCallback)] = [:]
+    fileprivate var publishRequests: [Int64: (callback: PublishCallback, errorCallback: ErrorPublishCallback)] = [:]
 
     // MARK: C'tor
     required public init(realm: String, transport: SwampTransport, authmethods: [String]?=nil, authid: String?=nil, authrole: String?=nil, authextra: [String: Any]?=nil){
@@ -154,7 +154,7 @@ open class SwampSession: SwampTransportDelegate {
     }
 
     // Internal because only a Subscription object can call this
-    internal func unsubscribe(_ subscription: Int, onSuccess: @escaping UnsubscribeCallback, onError: @escaping ErrorUnsubscribeCallback) {
+    internal func unsubscribe(_ subscription: Int64, onSuccess: @escaping UnsubscribeCallback, onError: @escaping ErrorUnsubscribeCallback) {
         let unsubscribeRequestId = self.generateRequestId()
         // Tell router to unsubscribe me from some subscription
         self.sendMessage(UnsubscribeSwampMessage(requestId: unsubscribeRequestId, subscription: subscription))
@@ -357,7 +357,7 @@ open class SwampSession: SwampTransportDelegate {
         self.transport.sendData(data)
     }
 
-    fileprivate func generateRequestId() -> Int {
+    fileprivate func generateRequestId() -> Int64 {
         self.currRequestId += 1
         return self.currRequestId
     }
